+++ logictree/nodes.py
@@ -7,6 +7,10 @@ class LogicNode:
     def depth(self):
         raise NotImplementedError
 
+    def inputs(self) -> set[str]:
+        return self.collect_input_names()
+        #raise NotImplementedError
+
     def to_verilog(self):
         raise NotImplementedError
 
@@ -15,11 +19,17 @@ class LogicNode:
 
     def contains_hole(self):
         return False
+    
+    def collect_input_names(self) -> set["LogicVar"]:
+        return NotImplementedError("Subclasses should implement collect_inputs")
 
 
 class LogicVar(LogicNode):
     def __init__(self, name):
         self.name = name
+
+    def inputs(self) -> set[str]:
+        return self.collect_input_names()
 
     def depth(self):
         return 0
@@ -38,10 +48,16 @@ class LogicVar(LogicNode):
     def __repr__(self):
         return f"LogicVar({self.name})"
 
+    def collect_input_names(self) -> set[str]:
+        return {self.name}
+
 
 class LogicConst(LogicNode):
     def __init__(self, value):
         self.value = value
+
+    def inputs(self) -> set[str]:
+        return self.collect_input_names()
 
     def depth(self):
         return 0
@@ -60,10 +76,16 @@ class LogicConst(LogicNode):
     def __repr__(self):
         return f"LogicConst({self.value})"
 
+    def collect_input_names(self) -> set[str]:
+        return set()
+
 
 class LogicHole(LogicNode):
     def __init__(self, name="UNSPECIFIED"):
         self.name = name
+
+    def inputs(self) -> set[str]:
+        return self.collect_input_names()
 
     def depth(self):
         return 0
@@ -85,24 +107,32 @@ class LogicHole(LogicNode):
     def __repr__(self):
         return f"LogicHole({self.name})"
 
+    def collect_input_names(self) -> set[str]:
+        return {self.name}
+
 
 class LogicOp(LogicNode):
     def __init__(self, op, inputs):
         assert op in GATE_TYPES
+        assert isinstance(op, str)
         self.name = op
-        self.inputs = [
+        self.op = op
+        self.children = [
                 inp if isinstance(inp, LogicNode) else LogicHole(f"input_{i}")
                 for i, inp in enumerate(inputs)
         ]
 
+    def inputs(self) -> set[str]:
+        return self.collect_input_names()
+
     def depth(self):
-        return 1 + max(inp.depth() for inp in self.inputs)
+        return 1 + max(inp.depth() for inp in self.children)
 
     def to_verilog(self):
         if self.name == "NOT":
-            print(f"DEBUG: LogicOp::to_verilog() self.inputs[0]:{self.inputs[0]}, name: {self.name}")
-            return f"~({self.inputs[0].to_verilog()})"
-        return f"({self.inputs[0].to_verilog()} {self.name} {self.inputs[1].to_verilog()})"
+            print(f"DEBUG: LogicOp::to_verilog() self.children[0]:{self.children[0]}, name: {self.name}")
+            return f"~({self.children[0].to_verilog()})"
+        return f"({self.children[0].to_verilog()} {self.name} {self.children[1].to_verilog()})"
 
     def to_dot(self, graph, parent_id=None, next_id=[0]):
         my_id = next_id[0]
@@ -110,15 +140,21 @@ class LogicOp(LogicNode):
         graph.append(f'  node{my_id} [label="{self.name}", shape=circle];')
         if parent_id is not None:
             graph.append(f'  node{parent_id} -> node{my_id};')
-        for inp in self.inputs:
+        for inp in self.children:
             inp.to_dot(graph, my_id, next_id)
         return my_id
 
     def contains_hole(self):
-        return any(inp.contains_hole() for inp in self.inputs)
+        return any(inp.contains_hole() for inp in self.children)
 
     def __repr__(self):
-        return f"LogicOp({self.name}, {self.inputs})"
+        return f"LogicOp({self.name}, {self.children})"
+
+    def collect_input_names(self) -> set[str]:
+        names = set()
+        for c in self.children:
+            names.update(c.collect_input_names())
+        return names
 
 
 # Utility API